// src/stores/interfaces/store.interface.ts
import { Document } from 'mongoose';

export interface Store extends Document {
  readonly storeID: string;
  readonly storeName: string;
  readonly takeOutInStore: boolean;
  readonly shippingTimeInDays: number;
  readonly latitude: string;
  readonly longitude: string;
  readonly address1: string;
  readonly address2: string;
  readonly address3: string;
  readonly city: string;
  readonly district: string;
  readonly state: string;
  readonly type: string;
  readonly country: string;
  readonly postalCode: string;
  readonly telephoneNumber: string;
  readonly emailAddress: string;
}